fix(pipeline): no need to hash
We can just use id.

+++ api/pipeline/pipeline.go
@@ -11,6 +11,6 @@ type (
 	Pipeline struct {
 		Name string `json:"name,omitempty"`
 		Jobs []*Job `json:"jobs,omitempty"`
-		ID   uint64 `json:"id,omitempty"`
+		ID   uint32 `json:"id,omitempty"`
 	}
 )
